refactor: api 훅으로 관리
