+++ src/features/placeDetail/PlaceDetailContainer.tsx
@@ -1,58 +1,20 @@
 import PhotoStrip from './PhotoStrip';
 import PlaceInfo from './PlaceInfo';
-import { useEffect, useState } from 'react';
-import axios from 'axios';
+
 import { useParams } from 'react-router-dom';
 import ReviewStats from './ReviewStats';
 import ReviewCard from './place-review-card/ReviewCard';
 import ReviewWriteButton from './ReviewWriteButton';
 import MoreReviewButton from './MoreReviewButton';
-import type { DetailPlaceProps, Review } from '../../types/type';
-import { useNavigate } from 'react-router-dom';
-import { getPlaceDetails } from './apis/placeDetailApi';
-import { getPlaceReview } from './apis/reviewApi';
-import { toast } from 'react-toastify';
+import { usePlaceDetail } from '../../hooks/usePlaceDetail';
+import { usePlaceReview } from '../../hooks/usePlaceReview';
 
 const PlaceDetailContainer = () => {
-  const [place, setPlace] = useState<DetailPlaceProps | null>(null);
-  const [reviews, setReviews] = useState<Review[] | []>([]);
   const { id } = useParams<{ id: string }>();
-  const navigate = useNavigate();
+  const { place } = usePlaceDetail(id!);
+  const { reviews } = usePlaceReview(id!);
 
-  useEffect(() => {
-    const fetchPlaceDetail = async () => {
-      try {
-        const placeData = await getPlaceDetails(id!);
-        setPlace(placeData);
-      } catch (err) {
-        if (axios.isAxiosError(err) && err.response) {
-          toast.error(err.response.data.message);
-          navigate(-1);
-        } else {
-          console.error(err);
-        }
-      }
-    };
-    fetchPlaceDetail();
-  }, [id, navigate]);
-
-  useEffect(() => {
-    const fetchPlaceReview = async () => {
-      try {
-        const reviewData = await getPlaceReview(id!);
-        setReviews(reviewData);
-      } catch (err) {
-        if (axios.isAxiosError(err) && err.response) {
-          toast.error(err.response.data.message);
-        } else {
-          console.error(err);
-        }
-      }
-    };
-    fetchPlaceReview();
-  }, [id]);
-
-  if (!place) return <div>로딩중...</div>;
+  if (!place || reviews.length === 0) return <div>로딩중...</div>;
 
   return (
     <div className="mx-auto mt-12 flex w-[75%] flex-col items-center gap-10 overflow-y-auto">
