+++ src/features/placeDetail/ReviewStats.tsx
@@ -1,29 +1,9 @@
 import { useParams } from 'react-router-dom';
-import { useState, useEffect } from 'react';
-import axios from 'axios';
-import type { ReviewStatsProps } from '../../types/type';
-import { getReviewStats } from './apis/reviewApi';
-import { toast } from 'react-toastify';
+import { useReviewStats } from '../../hooks/useReviewStats';
 
 const ReviewStats = () => {
   const { id } = useParams<{ id: string }>();
-  const [stats, setStats] = useState<ReviewStatsProps | null>(null);
-
-  useEffect(() => {
-    const fetchPlaceReviewStats = async () => {
-      try {
-        const statsData = await getReviewStats(id!);
-        setStats(statsData);
-      } catch (err) {
-        if (axios.isAxiosError(err) && err.response) {
-          toast.error(err.response.data.message);
-        } else {
-          console.error(err);
-        }
-      }
-    };
-    fetchPlaceReviewStats();
-  }, [id]);
+  const { stats } = useReviewStats(id!);
 
   if (!stats) {
     return <div> 리뷰 통계 로딩 중...</div>;
